backend/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
pydantic>=2.6
email-validator>=2.1
python-dotenv>=1.0
